Added preferences
Created a Preference struct for use later. These can be saved
into the flash memory so they will be preserved across reboots.

Also some code cleanup: refactor get_temperature since that
code was duplicated, improve the debugging messages, and change
everything to use farenheit instead of celsius.

Still left: second temp sensor (code is ready, but to connect
both devices to the power bus is a bit more soldering and wiring).
And the relay logic itself, which should compare the two temps,
maybe allowing overrides, and deal with the hysteresis.

Would be cool to have the web interface serve both json or html
depending on the Accept header, so if you use a browser you get
a nice pretty interface and if you use curl or some program you
get the parsable json.